Combat is now working with the wolf. His orientation is still messed up.

+++ Assets/Health.cs
@@ -6,13 +6,13 @@ public class Health : MonoBehaviour {
 	/// <summary>
 	/// Amount of health.
 	/// </summary>
-	public int healthValue = 100;
+	public float healthValue = 100.0f;
 
 	/// <summary>
 	/// Gets damaged by the input amount.
 	/// </summary>
 	/// <param name="dmg">Dmg.</param>
-	public void Damage(int dmg)
+	public void Damage(float dmg)
 	{
 		healthValue -= dmg;
 		Debug.Log ("I am now at " +  healthValue + " HP!");
